requirements.txt
configparser==5.3.0
paho-mqtt==1.6.1
pyserial==3.5